[HTTPError] Refactor HTTPError
* convert from js to ts
* renamed file to match object naming conventions
* remove default export
