+++ src/schema/show.ts
@@ -7,7 +7,7 @@ import {
   existyValue,
   convertConnectionArgsToGravityArgs,
 } from "lib/helpers"
-import HTTPError from "lib/http_error"
+import { HTTPError } from "lib/HTTPError"
 import numeral from "./fields/numeral"
 import { exhibitionPeriod, exhibitionStatus } from "lib/date"
 import cached from "./fields/cached"
